backend/pyproject.toml
[project]
name = "eco-test2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beanie>=2.0.1",
    "fastapi>=0.135.1",
    "fastapi-users[beanie]>=15.0.4",
    "motor>=3.7.1",
    "pydantic[email]>=2.12.5",
    "uvicorn[standard]>=0.41.0",
    "dnspython>=2.6.1",
    "python-dotenv>=1.2.2",
    "pymongo>=4.16.0",
    "httpx>=0.28.1",
    "scikit-learn>=1.8.0",
    "joblib>=1.5.3",
    "pandas>=3.0.1",
    "numpy>=2.4.3",
]
